feat: setup summarize route
The /summarize endpoint receives a json payload with the search query
and the number of max results. The function invoked when an HTTP
requests comes to this endpoint reads the json data and calls the
controller that sends back the summaries of each video that appeared in
the search results.